requirements.txt
fastmcp>=0.1.0
python-dotenv>=1.0.0
pocketflow>=0.1.0
google-generativeai>=0.3.0
asyncio
typing 